v1.7.7
trop.sh:
 - Style changes.
 - Multiple trop_mtl_common() fixes, and by extension,
   mv_torrent_location[_tracker]() bug fixes.
 - Added the global var `tret' to allow die() to exit with return
   value 0.
 - Various optimizations.

trop.awk:
 - Make the assert message a bit more specific in tracker_total().

install.sh:
 - Add -f switch.

README*:
 - Document install.sh addition.

Vim modelines were added. This was the only addition in tropriv.sh
so it's hash wasn't updated.

+++ install.sh
@@ -5,18 +5,34 @@ set -e
 
 err ()
 {
+	if [ $force -eq 1 ]; then
+		set +e
+		echo -e "error (ignored):" "$@"
+		return 0
+	fi
 	[ -n "$tmpfile" ] && rm ${tmpfile}
 	echo -e "install.sh stopped - error:" "$@" ; exit 1
 }
 
 trap 'echo ; err "caught signal"' SIGINT
 : ${PREFIX:=${HOME}/.trop}
+force=0
 
 if [ -n "$1" ]; then
+	usage='usage: install.sh [-f] [up]|[update]'
+	notupdate=0
 	case $1 in
-	up|update) break                                   ;;
-	*) echo 'usage: install.sh [up]|[update]' ; exit 1 ;;
+	up|update) break ;;
+	-f) force=1
+		if [ -z "$2" ]; then
+			notupdate=1 ; break
+		elif [ "$2" != 'up' ] && [ "$2" != 'update' ]; then
+			echo $usage ; exit 1
+		fi
+		;;
+	*) echo $usage ; exit 1 ;;
 	esac
+	[ $notupdate -eq 1 ] && break # -f was used for install
 	[ ! -e ${PREFIX} ] && err "\`$PREFIX' doesn't exist"
 	[ ! -d ${PREFIX} ] && err "PREFIX \`$PREFIX' is not a directory!"
 	[ ! -e ${PREFIX}/.is_trop_dir ] && \
@@ -109,3 +125,4 @@ fi
 
 echo "trop installed to \`${PREFIX}' successfully"
 exit 0
+# vim: ft=sh:ts=4:sw=4
